Modifying return type for last packet SNR to int8_t from int16_t

+++ include/spectre/driver/lora/RFM9x/RFM9x_Status.h
@@ -58,7 +58,7 @@ public:
 
   virtual int16_t getCurrentRssi   () override;
   virtual int16_t getLastPacketRssi() override;
-  virtual int16_t getLastPacketSnr () override;
+  virtual int8_t  getLastPacketSnr () override;
 
 
 private:
